remove code using 2nd core (was disabled anyway) reduce complexity in (ab)use of linker and symbols move `arm_math_m0` to a module (note: it is broken)

+++ users/elpekenin/src/sipo.c
@@ -35,7 +35,7 @@ static inline void print_sipo_byte(uint8_t x) {
 static inline void print_sipo_status(void) {
     logging(SIPO, LOG_DEBUG, "MCU");
 
-    for (int i = (SIPO_BYTES - 1); i >= 0; --i) {
+    for (uint8_t i = (SIPO_BYTES - 1); i >= 0; --i) {
         print_sipo_byte(i);
     }
